Move common modules from cli to api
