+++ src/snowcli/app/commands_registration/exception_logging.py
