+++ src/snowcli/cli/snowpark/commands.py
@@ -19,6 +19,7 @@ from snowcli.api.commands.flags import (
     DEFAULT_CONTEXT_SETTINGS,
     execution_identifier_argument,
 )
+from snowcli.api.constants import DEPLOYMENT_STAGE, ObjectType
 from snowcli.api.exception import (
     SecretsWithoutExternalAccessIntegrationError,
 )
@@ -32,7 +33,6 @@ from snowcli.api.output.types import (
 from snowcli.api.project_initialisation import add_init_command
 from snowcli.api.utils import file_utils
 from snowcli.api.utils.zipper import add_file_to_existing_zip
-from snowcli.cli.constants import DEPLOYMENT_STAGE, ObjectType
 from snowcli.cli.object.manager import ObjectManager
 from snowcli.cli.object.stage.manager import StageManager
 from snowcli.cli.snowpark.common import (
