+++ src/snowcli/cli/snowpark/common.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import os
 from typing import Dict, List, Optional
 
+from snowcli.api.constants import ObjectType
 from snowcli.api.sql_execution import SqlExecutionMixin
-from snowcli.cli.constants import ObjectType
 from snowcli.cli.snowpark.package_utils import generate_deploy_stage_name
 from snowflake.connector.cursor import SnowflakeCursor
