+++ src/snowcli/cli/snowpark/package/manager.py
@@ -7,9 +7,9 @@ from pathlib import Path
 from shutil import rmtree
 
 from requirements.requirement import Requirement
+from snowcli.api.constants import PACKAGES_DIR
 from snowcli.api.utils import file_utils
 from snowcli.api.utils.zipper import zip_dir
-from snowcli.cli.constants import PACKAGES_DIR
 from snowcli.cli.object.stage.manager import StageManager
 from snowcli.cli.snowpark import package_utils
 from snowcli.cli.snowpark.models import SplitRequirements
