+++ tests/conftest.py
@@ -5,7 +5,7 @@ import logging
 import pytest
 from snowcli.api.cli_global_context import cli_context_manager
 from snowcli.api.config import config_init
-from snowcli.cli import loggers
+from snowcli.app import loggers
 
 pytest_plugins = ["tests.testing_utils.fixtures", "tests.project.fixtures"]
